:sparkles: [BE] 퀘스트 클리어 로그 저장 개발완료
- 퀘스트 클리어 로그 기록 하는 API 개발 완료했습니다

+++ api-server/LogServer/src/main/java/com/exodia/logserver/db/entity/GameInfo.java
@@ -30,5 +30,5 @@ public class GameInfo {
 	private LocalDateTime endTime;
 	@Enumerated(EnumType.STRING)
 	private GameStatus gameStatus;
-	private ArrayList<Integer> users;
+	private ArrayList<Long> users;
 }
